avoid closing over values in string eval when comparing values
When comparing values using Test::Builder::cmp_ok, a string eval is
used to match the caller location, and to inject an arbitrary comparison
op. When used with Devel::Cover, the string eval may close over any
variables used in it, and maintain them for the lifetime of the program.

Instead of comparing the values directly in the string eval, generate an
anonymous sub that will perform the comparison, and call it in a block
eval.

This is a reoccurrance of a previous problem that was fixed with
4fe707c2dc3fe4e80767be33ce41a2a65d082574. Some combination of perl or
module updates has brought back the issue. While there was a test
included with the bug, it only fails when run with Devel::Cover, which
isn't done during the automated or release testing.