move try_sig_mask to separate file to avoid always loading POSIX
POSIX is an extra module to load which many test scripts don't need
themselves. Test2::Util was loading it for the try_sig_mask function,
but this is only used in one place for the IPC mechanism. If the script
isn't loading Test2::IPC or running with threads, it is not needed.

Move the try_sig_mask function to a separate module, and use the new
module in Test2::IPC::Driver::Files. For backwards compatibility,
maintain a try_sig_mask function in Test2::Util, but have it load and
call the function in the new module. A new module is used rather than
just delay loading POSIX, because for something involving threads or
forks, loading at process start is preferrable.