Correct documented, but mispelled names in PODs
Accidentally stumbled upon this (Test2::Hub, follow_up()).  This is
clearly a typo... Not sure if there are more typos like this though.

Future me here!  Turned out looking for more typos is a lot
harder than I imagined.  So I just ran Pod::Coverage and checked
that all uncovered subroutines are really undocumented and not
misspelled.  This led to a few interesting discoveries that I will
note here but will otherwise not commit further to this branch.

   - Test2::Compare::Meta, Test2::Compare::EventMeta - I was trying
     to understand add_prop() since DEFINED CHECKS seem to be
     referring the get_prop_*() getters.  (They're not really
     methods per se, I know.)  They were a little confusing and
     I would (personally) make it clearer what the checks are and
     document add_prop() somewhere; but it's internal... I get it
     if it's not the top priority.

   - Test2::Require: this side-effect condition seems redundant?

 63         if (HAVE_WIDGETS()) {
 64             # We have widgets, do not skip
 65             return undef;
 66         }
 67         else {
 68             # No widgets, skip the test
 69             return "Skipped because there are no widgets"
                                          unless HAVE_WIDGETS();
 70         }                         ### ^^^^^^^^^^^^^^^^^^^^^ ?

   - Test2::Workflow: unrelated to POD fix, but this shouldn't be a period?

 30                 die "Code is already set, did you provide multiple "
                        . "code blocks at $caller->[1] line $caller->[2].\n"
 31                     if $out{code};                              ### ^ ?

     and there is another one...

 46         die "Name is already set to '$out{name}', cannot "
                . "set to '$arg', did you specify multiple names "
                . "at $caller->[1] line $caller->[2].\n"
 47             if $out{name};                  ### ^ ?

This took me a day (see commit date), but I'm glad I can now put this
down while being able to say with some confidence that there are no more
documented but misspelled names. :)